\docType{data}
\name{alpe_d_huez}
\alias{alpe_d_huez}
\alias{alpe_d_huez2}
\title{Alpe d'Huez}
\format{A data frame with the following columns.
\describe{
\item{Time}{Character time of ride in the form M' S".}
\item{NumericTime}{Numeric time of ride in minutes.}
\item{Name}{Name of rider.}
\item{Year}{Year of race.}
\item{Nationality}{Nationality of rider.}
\item{DrugUse}{Have allegations of drug use been made against the
rider?  In \code{alpe_d_huez} the values are "Y" or "N"; in
\code{alpe_d_huez2} this is a logical vector.}
}}
\description{
  Fastest times for the Alpe d'Huez stage of the Tour de
  France cycle race, plus some contextual information on
  year and drug use allegations.
}
\note{
  The data was kindly compiled by William Hogan.

  The dataset is not guaranteed to be error free.  Please
  double check the data if you want to use it for something
  serious.
}
\references{
  William Hogan \email{william.hogan@amd.com} compiled the
  data from
  \url{http://en.wikipedia.org/wiki/Alpe_d\%27Huez}.
  Richard Cotton \email{richierocks@gmail.com} made some
  modifications while importing it into R.
}
\keyword{datasets}

