% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningR-package.R
\docType{data}
\name{btw2017}
\alias{btw2017}
\title{Wahlkreisergebnisse 2017.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 316 rows and 47 columns.
}
\usage{
btw2017
}
\description{
Wahlkreisergebnisse 2017.
}
\keyword{datasets}
