% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningR-package.R
\docType{data}
\name{coronastats}
\alias{coronastats}
\title{Corona-Faelle in Duisburg.}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 405 rows and 5 columns.
}
\usage{
coronastats
}
\description{
Corona-Faelle in Duisburg.
}
\keyword{datasets}
