\docType{data}
\name{english_monarchs}
\alias{english_monarchs}
\title{English Monarchs}
\format{A data frame with the following columns.
\describe{
\item{name}{Name of monarch(s).}
\item{house}{Royal house of monarch(s).}
\item{start.of.reign}{Year they rose to power.}
\item{end.of.reign}{Year they left power.}
\item{domain}{Region of England ruled over.}
}}
\description{
  Names, dates and houses of English kings and queens from
  post-Roman rule (the fifth century) until England invaded
  Ireland in the Early 13th century.
}
\note{
  This dataset is a bit messy and ambiguous in places,
  because history is like that. In fact, the messy parts of
  the dataset are in general a good indicator that
  soemthing interesting was happening at the time. (See,
  for example, missing or multiple rulers, starts and ends
  of reigns in the same year, and rulers that appear
  several times with different territories.)  Evan defining
  a monarch of England is tricky. Most of the monarchs in
  this dataset were around before England existed (it
  consisted of seven territories called the heptarchy). The
  data stops before John I (the bad guy from the Robin Hood
  stories) because he proclaimed himself King of Ireland,
  although some people consider monarchs up to Anne, five
  hundred years later, to be English Monarchs even though
  they ruled over Ireland, Wales and Scotland to varying
  degrees.

  The heptarchy consisted of East Anglia, Essex, Kent,
  Mercia, Northumbria, Sussex and Wessex.  Northubria was
  originally divided into Deria and Bernicia.  There are
  also periods of Norse and Danish rule.

  The dataset was compiled from Wikipedia and thus is not
  guaranteed to be error free.  Please double check the
  data if you want to use it for something serious.
}
\references{
  Richard Cotton \email{richierocks@gmail.com} compiled the
  dataset from various Wikipedia pages.
  \url{http://en.wikipedia.org/wiki/Kings_of_england}
  \url{http://en.wikipedia.org/wiki/Kings_of_East_Anglia}
  \url{http://en.wikipedia.org/wiki/Kings_of_Essex}
  \url{http://en.wikipedia.org/wiki/Kings_of_Kent}
  \url{http://en.wikipedia.org/wiki/Kings_of_Mercia}
  \url{http://en.wikipedia.org/wiki/Kings_of_Northumbria}
  \url{http://en.wikipedia.org/wiki/Kings_of_Sussex}
  \url{http://en.wikipedia.org/wiki/Kings_of_Wessex}
}
\keyword{datasets}

