\docType{data}
\name{gonorrhoea}
\alias{gonorrhoea}
\title{Gonorrhoea}
\format{A data frame with the following columns.
\describe{
\item{Year}{Year that infected people visited the clinic.}
\item{Age.Group}{Age group of person infected.}
\item{Ethnicity}{Ethnicity of person infected.}
\item{Gender}{Gender of person infected.}
\item{Rate}{Number of infections per 100000 people.}
}}
\description{
  Rates of gonorrhoea infection in the US by year, with
  contextual information about age, ethnicity and gender.
}
\references{
  Compiled by Richard Cotton \email{richierocks@gmail.com}
  from \url{http://www.cdc.gov/std/stats11/tables/22b.htm}
}
\keyword{datasets}

