\docType{package}
\name{learningr}
\alias{learningr}
\alias{learningr-package}
\title{Datasets to accompany the book Learning R}
\description{
  \code{learningr} contains datasets that are used in
  examples in the book ``Learning R''.
}
\author{
  Richard Cotton \email{richierocks@gmail.com}
}
\references{
  Cotton, R (2013) \emph{Learning R} O Reilly.  ISBN
  978-1-4493-5710-8.
}

