% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pres2017.R
\docType{data}
\name{pres2017}
\alias{pres2017}
\title{Results of the first turn of the French 2017 presidential election in Paris voting booths}
\format{
A data frame with 896 rows (one per voting booth) and 24 variables:
* Code du département
* Libellé du département
* Code de la circonscription
* Libellé de la circonscription
* Code de la commune
* Libellé de la commune
* Code du b.vote
* Inscrits
* Abstentions
* Votants
* Blancs
* Nuls
* Exprimés
* ARTHAUD Nathalie
* ASSELINEAU François
* CHEMINADE Jacques
* DUPONT-AIGNAN Nicolas
* FILLON François
* HAMON Benoît
* LASSALLE Jean
* LE PEN Marine
* MACRON Emmanuel
* MÉLENCHON Jean-Luc
* POUTOU Philippe
}
\usage{
pres2017
}
\description{
A dataset containing the turnout and results of the first turn of the French 2017 presidential election in Paris voting booths
}
\keyword{datasets}
