% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningR-package.R
\docType{data}
\name{senti_ws}
\alias{senti_ws}
\title{SentiWS.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 34595 rows and 6 columns.
}
\usage{
senti_ws
}
\description{
SentiWS.
}
\keyword{datasets}
