% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learningR-package.R
\docType{data}
\name{corona_by_county}
\alias{corona_by_county}
\title{Corona auf Ebene der Landkreise.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 10 columns.
}
\usage{
corona_by_county
}
\description{
Corona auf Ebene der Landkreise.
}
\keyword{datasets}
